Remove use of Fortran 2008 standard to represent double-precision floating-point format (closes #2) (#3)
Removes the use of Fortran 2008 standard to represent double-precision floating-point to make the library more compatible with other software. By default, this library is now compiled enforcing real type to an 8 byte wide by passing the relevant flat to the compiler however the behaviour can be modified by the use of a new flag in CMake (USE_REAL8).